Subscribe for immediate updates
This removes the updateInterval polling approach and switches to a push-based approach where the equipment tells us when something changes and we show it immediately. Overall this is a reduction/simplification in code, but I do have concerns for the longer term that the connection will get closed and end up in a non-updating state. I think I've handled the main cases where this can happen (error and socket close events), but will need to run the module for a while to see what happens over a longer period of time.

While I was at it, I worked to ensure that additional clients connecting to the same Magic Mirror server don't put any additional load on the pool equipment itself; all pool status data is cached and simply passed to the new client instead of the previous method of triggering more connections and messages per client.

Oh, and ditched the pre-es6 workarounds now that I no longer need to support the old Midori browser for this module.